requirements.txt
numpy==1.26.1
Pillow==10.3.0
huggingface_hub
safetensors
roma
gradio
matplotlib
tqdm
opencv-python
opencv-python-headless
open_clip_torch
scipy
einops
trimesh
tensorboard
pyglet<2
viser
lpips
hydra-core
h5py
accelerate
transformers
scikit-learn
gsplat
evo
open3d
wandb
imageio
imageio-ffmpeg